Split property.py into seperate files under property directory
